LATX, feat: introduce smc strategy
The SMC (self-modify-code) is a tough problem for a binary translator. The
consistency between the guest program's binary codes and the generated host
binary codes should be correctly and efficiently managed.

This commit introduces different SMC strategy to be performed in LAT. Configure
it with an input option latx-smc or environment virable LAT_SMC. The default
value is 0.

On default, the SMC strategy is write-first. The page that have been translated
will be unwritable to capture the following store inst. In the segv handler,
all the TB on this page will be invalidated and make the page writable again.
Then the store inst can execute without triger segv fault.

On input value != 0, the SMC strategy is code-first. The page is also unwritable
after translation. In the segv handler, only the TB that is actually affected
by the store inst will be invalidated. The page can not become writable if there
are other TB on this page. Then the store inst is interpreted using the shadow
page mechanism, which allows the store inst to finish the write operation.

The write-first strategy could lead to a lot of TB being retranslated since it
will invalidate all the TB on one page no matter they are actually modified.
On the other hand, the code-first strategy could lead to a lot of SEGV fault
since the page keeps unwritable if there are TB on it. The balance between
these two overhead is the key to optimize the SMC strategy.

More optimizations could be done later:
- record the TB's guest codes to avoid retranslation if it is not modified
- use shared mmap instead of shadow page to allow read operation
- use shared mmap to truly enable write on those pages without TB
- retranslate the store inst to avoid segv fault and finish the TB-invalidate
  work in a fast path with acceptable overhead
- use some kind of bitmap to maintain the dirty state of one page and
  let TB check it during its execution

Signed-off-by: NiuGenen <niugenen@qq.com>