LATX, opt: use shared map for smc
Configure with LAT_SMC=2 (bit[1] is 1) to enable shared map for SMC strategy.

The original code-first strategy (LAT_SMC != 0) use shadow page for SMC to interpret
the store inst. But shadow page will disable both read and write on the original
guest pages, which leads to a lot of unnecessary SEGV caused by read.

This patch uses shared map to replace the shadow page in SMC strategy. The shared
memory space is create with linux memfd. The original guest pages are allow to
read directly. The write is still disabled to capture the store inst that might
modify guest's binary codes.

For now, only when the original guest space is MAP_ANON, can we use the shared map
strategy on it. If the guest space is already mmaped with a fd, then we could and
must use this fd to create shared map, instead of the memfd.

The ShadowPageDesc is used to manage the shared mapping. The multiple guest pages
belonging to the same host page are configured with the same p_addr and access_off.
If the original guest space is already shadowed, then no more shared map or shadow
page will be created.

Signed-off-by: NiuGenen <niugenen@qq.com>